Move widget-related utility code out of utils.py, into widgets.py
