+++ lib/streamlit/elements/checkbox.py
@@ -16,7 +16,7 @@ from typing import cast
 
 import streamlit
 from streamlit.proto.Checkbox_pb2 import Checkbox as CheckboxProto
-from .utils import register_widget
+from streamlit.widgets import register_widget
 
 
 class CheckboxMixin:
