+++ lib/streamlit/elements/multiselect.py
@@ -18,7 +18,7 @@ import streamlit
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.MultiSelect_pb2 import MultiSelect as MultiSelectProto
 from streamlit.type_util import is_type, ensure_iterable
-from .utils import register_widget
+from streamlit.widgets import register_widget
 
 
 class MultiSelectMixin:
