+++ lib/streamlit/elements/select_slider.py
@@ -18,7 +18,7 @@ import streamlit
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.Slider_pb2 import Slider as SliderProto
 from streamlit.type_util import ensure_iterable
-from .utils import register_widget
+from streamlit.widgets import register_widget
 
 
 class SelectSliderMixin:
