+++ lib/streamlit/elements/time_widgets.py
@@ -19,7 +19,7 @@ import streamlit
 from streamlit.errors import StreamlitAPIException
 from streamlit.proto.DateInput_pb2 import DateInput as DateInputProto
 from streamlit.proto.TimeInput_pb2 import TimeInput as TimeInputProto
-from .utils import register_widget
+from streamlit.widgets import register_widget
 
 
 class TimeWidgetsMixin:
